back/requirements.txt
# Firebase and Google Cloud
firebase-admin==6.5.0
firebase-functions==0.4.2
google-cloud-firestore==2.16.0
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.20.0

# Google APIs (Calendar integration)
google-api-python-client==2.142.0
google-auth==2.30.0
google-auth-oauthlib==1.2.0

# Web framework
Flask==3.0.3
flask-cors==4.0.1

# Data validation and serialization
pydantic==2.7.1
python-dateutil==2.9.0

# HTTP and API clients
requests==2.32.3
httpx==0.27.0

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0
pytest-mock==3.14.0
functions-framework==3.8.1

# Development tools
black==24.4.2
mypy==1.10.0
pylint==3.2.2

# Utilities
python-dotenv==1.0.1
pytz==2024.1